requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
idna==2.8
importlib-metadata==0.18
more-itertools==7.2.0
packaging==19.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
pytest-html==1.21.1
pytest-metadata==1.8.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
